requirements.txt
absl-py==0.13.0
anyio==3.1.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bleach==3.3.0
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.2.2
certifi==2022.6.15
cffi==1.14.5
charset-normalizer==2.1.0
cleo==0.8.1
click==8.1.3
clikit==0.6.2
colorama==0.4.4
crashtest==0.3.1
cryptography==35.0.0
cycler==0.10.0
dataclasses==0.6
dcmrtstruct2nii==1.0.19
debugpy==1.4.1
decorator==4.4.2
defusedxml==0.7.1
dicom-numpy==0.5.0
dicompyler-core==0.5.5
docker-pycreds==0.4.0
entrypoints==0.3
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==1.34.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.3
imageio==2.9.0
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
monai==0.8.1
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
neurite==0.1
nibabel==3.2.1
notebook==6.4.0
numexpr==2.7.3
numpy==1.19.5
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pastel==0.2.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pip==21.2.4
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydicom==2.1.2
Pygments==2.9.0
pylev==1.4.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pystrum==0.1
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.28.1
requests-oauthlib==1.3.0
rsa==4.7.2
rt-utils==1.1.4
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.6.3
Send2Trash==1.5.0
sentry-sdk==1.6.0
setproctitle==1.2.2
setuptools==58.0.4
shortuuid==1.0.9
SimpleITK==2.0.2
six==1.16.0
smmap==3.0.5
sniffio==1.2.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.7.2
torch==1.9.1
torchaudio==0.9.0a0+a85b239
torchvision==0.10.1
tornado==6.1
tqdm==4.63.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.9
voxelmorph==0.1
wandb==0.12.20
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
wheel==0.37.0
widgetsnbextension==3.5.2
wrapt==1.12.1
zipp==3.6.0
